chore(apps/pano): move funcs to utils file
